refactor: extract nesteds
